requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
asyncpg
psycopg[binary]
alembic
pinecone
redis
hiredis
langchain
langgraph
langsmith
langchain-community
requests
httpx
pillow
opencv-python
fpdf2
python-dotenv
python-multipart
websockets
langchain-groq
twilio
huggingface_hub
pytest
pytest-asyncio
mcp[cli]>=0.1.0